Create/update tasks, authorization

+++ src/main/java/com/TaskMate/TaskMate/repo/UserRepo.java
@@ -5,7 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepo extends JpaRepository<Users, Integer> {
-
+public interface UserRepo extends JpaRepository<Users, Long> {
     Users findByUsername(String username);
 }
